third_party/blink/web_tests/external/wpt/custom-elements/registries/scoped-custom-element-registry-customelementregistry-attribute-in-xhtml-expected.txt

This is a testharness.js-based test.
[FAIL] XHTML parser should create a custom element candidate with null registry if customelementregistry is set
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] XHTML parser should create a defined custom element with null registry if customelementregistry is set
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] XHTML fragment parser should create a custom element candidate with null registry if customelementregistry is set
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] XHTML fragment parser should create a defined custom element with null registry if customelementregistry is set
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
Harness: the test ran to completion.
